requirements.txt
blobfile==2.1.1
click==8.1.7
einops==0.7.0
matplotlib==3.8.0
mpi4py==3.1.5
numpy==1.26.4
Pillow==10.3.0
piq==0.8.0
psutil==5.9.0
pyspng==0.1.1
Requests==2.31.0
scikit_learn==1.2.2
scipy==1.13.0
tensorboardX==2.6.2.2
tensorboardX==2.6.2.2
tensorflow==2.16.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.65.0
